+++ src/regparse.c
@@ -1868,7 +1868,7 @@ callout_tag_entry(regex_t* reg, UChar* name, UChar* name_end,
   ext = onig_get_regex_ext(reg);
   r = callout_tag_entry_raw(ext->tag_table, name, name_end, entry_val);
 
-  e = onig_reg_callout_list_at(reg, entry_val);
+  e = onig_reg_callout_list_at(reg, (int )entry_val);
   e->tag_start = name;
   e->tag_end   = name_end;
